proxy: Use file instead of string for vnc password
For security reasons, the password should not show up in the command
line. We replace the '--vnc-password' option with the
'--vnc-password-file' option that expects a file path as its value.

This file should contain the password in plain text in the first
line. The remaining lines will be ignored.

Update the docs accordingly.

Signed-off-by: Dimitris Aragiorgis <dimitris.aragiorgis@gmail.com>