proxy: Support multiple VNC passwords
Add support for authenticating against VNC servers that require
different passwords.

The VNC password file should now include one password per line,
where lines starting with "#" are ignored.

If the server requests for password authentication, then the proxy
will try to connect using available passwords in the order found in
the password file.

If a server changes its password then this password should be added
in the VNC password file and vncauthproxy should be restarted.

This patch improves logging with regard to the VNC authentication
progress.

Signed-off-by: Dimitris Aragiorgis <dimitris.aragiorgis@gmail.com>